refactor(admin): adopt shared RawJsonTab, drop the per-resource copies (#1104)
Eight resources each shipped their own tabs/raw-json-tab.tsx — six
byte-identical to the existing @/common/RawJsonTab, two differing only in
a cosmetic useRecordContext type param that is passed straight to
JsonOutput and changes nothing. log-streams carried a ninth inline copy.

Point all of them at @/common/RawJsonTab and delete the copies. No
behaviour change.
