--- apps/admin/src/resources/connections/tabs/raw-json-tab.tsx
@@ -1,8 +1,0 @@
-import { useRecordContext } from "ra-core";
-import { JsonOutput } from "@/common/JsonOutput";
-
-export function RawJsonTab() {
-  const record = useRecordContext();
-  if (!record) return null;
-  return <JsonOutput data={record} />;
-}
