--- apps/admin/src/resources/roles/tabs/raw-json-tab.tsx
@@ -1,9 +1,0 @@
-import { useRecordContext } from "ra-core";
-import type { RaRecord } from "ra-core";
-import { JsonOutput } from "@/common/JsonOutput";
-
-export function RawJsonTab() {
-  const record = useRecordContext<RaRecord>();
-  if (!record) return null;
-  return <JsonOutput data={record} />;
-}
