--- apps/admin/src/resources/users/tabs/raw-json-tab.tsx
@@ -1,9 +1,0 @@
-import { useRecordContext } from "ra-core";
-import { JsonOutput } from "@/common/JsonOutput";
-import type { UserRecord } from "./types";
-
-export function RawJsonTab() {
-  const record = useRecordContext<UserRecord>();
-  if (!record) return null;
-  return <JsonOutput data={record} />;
-}
